完成一个数据库查询demo

+++ springmvc-01/src/main/java/spitter/db/SpitterRepository.java
@@ -12,7 +12,7 @@ public interface SpitterRepository {
 
     Spitter save(Spitter spitter);
 
-    Spitter findOne(long id);
+    Spitter findOne(int id);
 
     Spitter findByUsername(String username);
